+++ packages/build-app-helpers/src/tests/fixtures/analyzeRuntime/page/index.js
@@ -1,0 +1,4 @@
+import { useAuth } from 'ice';
+import store from './store';
+
+export const page = () => {};