+++ packages/plugin-ice-ssr/src/env.ts
@@ -1,0 +1,2 @@
+// setup server env
+process.env.__IS_SERVER__ = 'true';