+++ packages/plugin-router/src/runtime/Router.tsx
@@ -106,7 +106,6 @@ export function Routes({ routes, fallback }: RoutesProps) {
     <Switch>
       {routes.map((route, id) => {
         const { children } = route;
-
         if (!children) {
           if (route.redirect) {
             const { redirect, ...others } = route;