+++ packages/react-app-renderer/src/server.tsx
@@ -16,18 +16,24 @@ function renderInServer(context: Context, options: RenderOptions) {
   emitLifeCycles();
 
   const App = getRenderApp(runtime, options);
-
-  const webExtractor = new ChunkExtractor({
-    statsFile: loadableStatsPath,
-    entrypoints: ['index'],
-    publicPath
-  });
-  const jsx = webExtractor.collectChunks(<App />);
-
-  return {
-    bundleContent: ReactDOMServer.renderToString(jsx),
-    loadableComponentExtractor: webExtractor
-  };
+  
+  if (loadableStatsPath) {
+    const webExtractor = new ChunkExtractor({
+      statsFile: loadableStatsPath,
+      entrypoints: ['index'],
+      publicPath
+    });
+    const jsx = webExtractor.collectChunks(<App />);
+  
+    return {
+      bundleContent: ReactDOMServer.renderToString(jsx),
+      loadableComponentExtractor: webExtractor
+    };
+  } else {
+    return {
+      bundleContent: ReactDOMServer.renderToString(<App />)
+    };
+  }
 }
 
 export default function reactAppRendererWithSSR(context: Context, options: RenderOptions) {
@@ -47,11 +53,17 @@ function deepClone(config) {
   if(typeof config !== 'object' || config === null) {
     return config;
   }
-  const ret = {};
-  Object.getOwnPropertyNames(config).forEach((key: string) => {
-    if (Object.prototype.hasOwnProperty.call(config, key)) {
-      ret[key] = deepClone(config[key]);
-    }
-  });
-  return ret;
+
+  if (Array.isArray(config)) {
+    return config.slice();
+  } else {
+    const ret = {};
+    Object.getOwnPropertyNames(config).forEach((key: string) => {
+      if (Object.prototype.hasOwnProperty.call(config, key)) {
+        ret[key] = deepClone(config[key]);
+      }
+    });
+    return ret;
+  }
+  
 }