+++ packages/vite-plugin-index-html/src/index.ts
@@ -1,8 +1,8 @@
-import { extname, resolve, posix } from 'path';
+import { extname, resolve, normalize, join, isAbsolute } from 'path';
 import { readFileSync, pathExists } from 'fs-extra';
 import type { Plugin, ResolvedConfig, HtmlTagDescriptor } from 'vite';
 import type { OutputBundle, OutputAsset, OutputChunk, PreserveEntrySignaturesOption } from 'rollup';
-import { isAbsoluteUrl, addTrailingSlash, formatPath, getEntryUrl, isSingleEntry } from './utils';
+import { isAbsoluteUrl, addTrailingSlash, formatPath, getEntryUrl, isSingleEntry, getRelativePath } from './utils';
 import minifyHtml from './minifyHtml';
 
 const scriptLooseRegex = /<script\s[^>]*src=['"]?([^'"]*)['"]?[^>]*>*<\/script>/;
@@ -134,6 +134,7 @@ export default function htmlPlugin(userOptions?: HtmlPluginOptions): Plugin {
 
       const html = injectToHtml(
         removeHtmlEntryScript(
+          viteConfig.root,
           parseTemplate(userOptions.template, userOptions.templateContent),
           getEntryUrl(userOptions.input),
         ),
@@ -155,7 +156,7 @@ export default function htmlPlugin(userOptions?: HtmlPluginOptions): Plugin {
         tag: 'script',
         attrs: {
           type: 'module',
-          src: getRelativedPath(
+          src: getRelativePath(
             viteConfig.root,
             getEntryUrl(userOptions.input)
           ),
@@ -164,6 +165,7 @@ export default function htmlPlugin(userOptions?: HtmlPluginOptions): Plugin {
       };
       const _html = injectToHtml(
         removeHtmlEntryScript(
+          viteConfig.root,
           html,
           getEntryUrl(userOptions.input)),
         [entryTag],
@@ -173,9 +175,12 @@ export default function htmlPlugin(userOptions?: HtmlPluginOptions): Plugin {
   };
 }
 
-export const removeHtmlEntryScript = (html: string, entry: string) => {
+export const removeHtmlEntryScript = (rootDir: string, html: string, entry: string) => {
   let _html = html;
-  const _entry = formatPath(entry);
+  const _entry = formatPath(
+    isAbsolute(entry) ? entry : join(rootDir, entry)
+  );
+
   const matchs = html.match(new RegExp(scriptLooseRegex, 'g'));
 
   const commentScript = (script: string) => `<!-- removed by vite-plugin-index-html ${script} -->`;
@@ -183,8 +188,10 @@ export const removeHtmlEntryScript = (html: string, entry: string) => {
   if (matchs) {
     matchs.forEach((matchStr) => {
       const [, src] = matchStr.match(scriptLooseRegex);
+      // change `./src/index.js` to `src/index.js`
+      const normalizedSrc = normalize(src);
 
-      if (_entry.includes(src)) {
+      if (_entry.includes(normalizedSrc)) {
         _html = _html.replace(matchStr, commentScript(matchStr));
         console.warn(`
           vite-plugin-index-html: For the reason that entry was configured, ${matchStr} is deleted.
@@ -248,15 +255,6 @@ function getFiles(bundle: OutputBundle): OutputBundleExt[] {
 
 function toPublicPath(filename: string, config: ResolvedConfig) {
   return isAbsoluteUrl(filename) ? filename : addTrailingSlash(config.base) + filename;
-}
-
-function getRelativedPath(rootDir: string, path: string): string {
-  const _rootDir = formatPath(rootDir);
-  let _path = formatPath(path);
-  if (path.includes(_rootDir)) {
-    _path = `/${posix.relative(_rootDir, _path)}`;
-  }
-  return _path;
 }
 
 const headInjectRE = /<\/head>/;