+++ packages/vite-service/src/build.ts
@@ -9,9 +9,14 @@ export async function viteBuild(context: any): Promise<BuildResult> {
   const { applyHook, command, commandArgs } = context;
 
   const configArr = context.getWebpackConfig();
-  await applyHook(`before.${command}.load`, { args: commandArgs, webpackConfig: configArr });
+  const prodConfig = configArr.length > 0 ? wp2vite(context) : {};
+  await applyHook(`before.${command}.load`, { args: commandArgs, webpackConfig: configArr, viteConfig: prodConfig });
 
-  const prodConfig = wp2vite(context);
+  if (!configArr.length) {
+    const errorMsg = 'No config found.';
+    await applyHook('error', { err: new Error(errorMsg) });
+    return;
+  }
 
   await applyHook(`before.${command}.run`, {
     args: commandArgs,
@@ -20,7 +25,10 @@ export async function viteBuild(context: any): Promise<BuildResult> {
 
   try {
     await build(prodConfig);
-    context.applyHook('after.build.compile');
+    context.applyHook('after.build.compile', {
+      args: commandArgs,
+      config: prodConfig,
+    });
   } catch (err) {
     console.error('CONFIG', chalk.red('Failed to load vite config.'));
     throw err;