+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.3
+
+- [fix] filter runtime modules after file generation
+
+## 2.1.2
+
+- [fix] add buildConfig of i18n
+
 ## 2.1.1
 
 - [fix] circular dependence of lazy when run test