+++ packages/vite-service/CHANGELOG.md
@@ -1,8 +1,15 @@
 # CHANGELOG
 
-## 2.0.4
+## 2.1.1
 
 - [feat] plugin for rename main css chunk name
+
+## 2.1.0
+
+- [feat] refactor html plugin for SSR render when development
+- [fix] lifecycle arguments for `after.build.compile`
+- [fix] pass vite config for lifecycle hooks
+- [fix] fail to open browser in mpa
 
 ## 2.0.3